giga cleanup (elijah said I should call it this)

+++ mojoPortal.Data.MySql/DBAuthorizeNetLog.cs
@@ -1,7 +1,6 @@
-﻿using MySqlConnector;
-using System;
-using System.Collections.Generic;
+﻿using System;
 using System.Data;
+using MySqlConnector;
 
 namespace mojoPortal.Data;
 
@@ -57,189 +56,80 @@ public static class DBAuthorizeNetLog
 	{
 		var sqlCommand = @"
 INSERT INTO mp_AuthorizeNetLog (
-	RowGuid, 
-	CreatedUtc, 
-	SiteGuid, 
-	UserGuid, 
-	StoreGuid, 
-	CartGuid, 
-	RawResponse, 
-	ResponseCode, 
-	ResponseReasonCode, 
-	Reason, 
-	AvsCode, 
-	CcvCode, 
-	CavCode, 
-	TransactionId, 
-	TransactionType, 
-	Method, 
-	AuthCode, 
-	Amount, 
-	Tax, 
-	Duty, 
-	Freight 
+	RowGuid
+	,CreatedUtc
+	,SiteGuid
+	,UserGuid
+	,StoreGuid
+	,CartGuid
+	,RawResponse
+	,ResponseCode
+	,ResponseReasonCode
+	,Reason
+	,AvsCode
+	,CcvCode
+	,CavCode
+	,TransactionId
+	,TransactionType
+	,Method
+	,AuthCode
+	,Amount
+	,Tax
+	,Duty
+	,Freight 
 )
 VALUES (
-	?RowGuid, 
-	?CreatedUtc, 
-	?SiteGuid, 
-	?UserGuid, 
-	?StoreGuid, 
-	?CartGuid, 
-	?RawResponse, 
-	?ResponseCode, 
-	?ResponseReasonCode, 
-	?Reason, 
-	?AvsCode, 
-	?CcvCode, 
-	?CavCode, 
-	?TransactionId, 
-	?TransactionType, 
-	?Method, 
-	?AuthCode, 
-	?Amount, 
-	?Tax, 
-	?Duty, 
-	?Freight 
+	?RowGuid
+	,?CreatedUtc
+	,?SiteGuid
+	,?UserGuid
+	,?StoreGuid
+	,?CartGuid
+	,?RawResponse
+	,?ResponseCode
+	,?ResponseReasonCode
+	,?Reason
+	,?AvsCode
+	,?CcvCode
+	,?CavCode
+	,?TransactionId
+	,?TransactionType
+	,?Method
+	,?AuthCode
+	,?Amount
+	,?Tax
+	,?Duty
+	,?Freight 
 );";
 
-		var arParams = new List<MySqlParameter>()
-		{ 
-			new("?RowGuid", MySqlDbType.VarChar, 36)
-				{
-					Direction = ParameterDirection.Input,
-					Value = rowGuid.ToString()
-				},
-
-			new("?CreatedUtc", MySqlDbType.DateTime)
-				{
-					Direction = ParameterDirection.Input,
-					Value = createdUtc
-				},
-
-			new("?SiteGuid", MySqlDbType.VarChar, 36)
-				{
-					Direction = ParameterDirection.Input,
-					Value = siteGuid.ToString()
-				},
-
-			new("?UserGuid", MySqlDbType.VarChar, 36)
-				{
-					Direction = ParameterDirection.Input,
-					Value = userGuid.ToString()
-				},
-
-			new("?StoreGuid", MySqlDbType.VarChar, 36)
-				{
-					Direction = ParameterDirection.Input,
-					Value = storeGuid.ToString()
-				},
-
-			new("?CartGuid", MySqlDbType.VarChar, 36)
-				{
-					Direction = ParameterDirection.Input,
-					Value = cartGuid.ToString()
-				},
-
-			new("?RawResponse", MySqlDbType.Text)
-				{
-					Direction = ParameterDirection.Input,
-					Value = rawResponse
-				},
-
-			new("?ResponseCode", MySqlDbType.VarChar, 1)
-				{
-					Direction = ParameterDirection.Input,
-					Value = responseCode
-				},
-
-			new("?ResponseReasonCode", MySqlDbType.VarChar, 20)
-				{
-					Direction = ParameterDirection.Input,
-					Value = responseReasonCode
-				},
-
-			new("?Reason", MySqlDbType.Text)
-				{
-					Direction = ParameterDirection.Input,
-					Value = reason
-				},
-
-				new("?AvsCode", MySqlDbType.VarChar, 50)
-				{
-					Direction = ParameterDirection.Input,
-					Value = avsCode
-				},
-
-				new("?CcvCode", MySqlDbType.VarChar, 1)
-				{
-					Direction = ParameterDirection.Input,
-					Value = ccvCode
-				},
-
-				new("?CavCode", MySqlDbType.VarChar, 1)
-				{
-					Direction = ParameterDirection.Input,
-					Value = cavCode
-				},
-
-				new("?TransactionId", MySqlDbType.VarChar, 50)
-				{
-					Direction = ParameterDirection.Input,
-					Value = transactionId
-				},
-
-				new("?TransactionType", MySqlDbType.VarChar, 50)
-				{
-					Direction = ParameterDirection.Input,
-					Value = transactionType
-				},
-
-				new("?Method", MySqlDbType.VarChar, 20)
-				{
-					Direction = ParameterDirection.Input,
-					Value = method
-				},
-
-				new("?AuthCode", MySqlDbType.VarChar, 50)
-				{
-					Direction = ParameterDirection.Input,
-					Value = authCode
-				},
-
-				new("?Amount", MySqlDbType.Decimal)
-				{
-					Direction = ParameterDirection.Input,
-					Value = amount
-				},
-
-				new("?Tax", MySqlDbType.Decimal)
-				{
-					Direction = ParameterDirection.Input,
-					Value = tax
-				},
-
-				new("?Duty", MySqlDbType.Decimal)
-				{
-					Direction = ParameterDirection.Input,
-					Value = duty
-				},
-
-				new("?Freight", MySqlDbType.Decimal)
-				{
-					Direction = ParameterDirection.Input,
-					Value = freight
-				}
+		var arParams = new MySqlParameter[]
+		{
+			new("?RowGuid", MySqlDbType.VarChar, 36) { Value = rowGuid },
+			new("?CreatedUtc", MySqlDbType.DateTime) { Value = createdUtc },
+			new("?SiteGuid", MySqlDbType.VarChar, 36) { Value = siteGuid },
+			new("?UserGuid", MySqlDbType.VarChar, 36) { Value = userGuid },
+			new("?StoreGuid", MySqlDbType.VarChar, 36) { Value = storeGuid },
+			new("?CartGuid", MySqlDbType.VarChar, 36) { Value = cartGuid },
+			new("?RawResponse", MySqlDbType.Text) { Value = rawResponse },
+			new("?ResponseCode", MySqlDbType.VarChar, 1) { Value = responseCode },
+			new("?ResponseReasonCode", MySqlDbType.VarChar, 20) { Value = responseReasonCode },
+			new("?Reason", MySqlDbType.Text) { Value = reason },
+			new("?AvsCode", MySqlDbType.VarChar, 50) { Value = avsCode },
+			new("?CcvCode", MySqlDbType.VarChar, 1) { Value = ccvCode },
+			new("?CavCode", MySqlDbType.VarChar, 1) { Value = cavCode },
+			new("?TransactionId", MySqlDbType.VarChar, 50) { Value = transactionId },
+			new("?TransactionType", MySqlDbType.VarChar, 50) { Value = transactionType },
+			new("?Method", MySqlDbType.VarChar, 20) { Value = method },
+			new("?AuthCode", MySqlDbType.VarChar, 50) { Value = authCode },
+			new("?Amount", MySqlDbType.Decimal) { Value = amount },
+			new("?Tax", MySqlDbType.Decimal) { Value = tax },
+			new("?Duty", MySqlDbType.Decimal) { Value = duty },
+			new("?Freight", MySqlDbType.Decimal) { Value = freight }
 		};
 
-		int rowsAffected = CommandHelper.ExecuteNonQuery(
-			ConnectionString.GetWriteConnectionString(),
-			sqlCommand.ToString(),
-			arParams);
+		int rowsAffected = CommandHelper.ExecuteNonQuery(ConnectionString.GetWrite(), sqlCommand, arParams);
 		return rowsAffected;
-
 	}
-
 
 	/// <summary>
 	/// Updates a row in the mp_AuthorizeNetLog table. Returns true if row updated.
@@ -292,139 +182,55 @@ VALUES (
 UPDATE 
 	mp_AuthorizeNetLog 
 SET  
-	SiteGuid = ?SiteGuid, 
-	UserGuid = ?UserGuid, 
-	StoreGuid = ?StoreGuid, 
-	CartGuid = ?CartGuid, 
-	RawResponse = ?RawResponse, 
-	ResponseCode = ?ResponseCode, 
-	ResponseReasonCode = ?ResponseReasonCode, 
-	Reason = ?Reason, 
-	AvsCode = ?AvsCode, 
-	CcvCode = ?CcvCode, 
-	CavCode = ?CavCode, 
-	TransactionId = ?TransactionId, 
-	TransactionType = ?TransactionType, 
-	Method = ?Method, 
-	AuthCode = ?AuthCode, 
-	Amount = ?Amount, 
-	Tax = ?Tax, 
-	Duty = ?Duty, 
-	Freight = ?Freight 
+	SiteGuid = ?SiteGuid 
+	,UserGuid = ?UserGuid 
+	,StoreGuid = ?StoreGuid 
+	,CartGuid = ?CartGuid 
+	,RawResponse = ?RawResponse 
+	,ResponseCode = ?ResponseCode 
+	,ResponseReasonCode = ?ResponseReasonCode 
+	,Reason = ?Reason 
+	,AvsCode = ?AvsCode 
+	,CcvCode = ?CcvCode 
+	,CavCode = ?CavCode 
+	,TransactionId = ?TransactionId 
+	,TransactionType = ?TransactionType 
+	,Method = ?Method 
+	,AuthCode = ?AuthCode 
+	,Amount = ?Amount 
+	,Tax = ?Tax 
+	,Duty = ?Duty 
+	,Freight = ?Freight 
 WHERE  
 	RowGuid = ?RowGuid;";
 
-		var arParams = new List<MySqlParameter>
+		var arParams = new MySqlParameter[]
 		{
-			new("?RowGuid", MySqlDbType.VarChar, 36)
-			{
-				Direction = ParameterDirection.Input,
-				Value = rowGuid.ToString()
-			},
-			new("?SiteGuid", MySqlDbType.VarChar, 36)
-			{
-				Direction = ParameterDirection.Input,
-				Value = siteGuid.ToString()
-			},
-			new("?UserGuid", MySqlDbType.VarChar, 36)
-			{
-				Direction = ParameterDirection.Input,
-				Value = userGuid.ToString()
-			},
-			new("?StoreGuid", MySqlDbType.VarChar, 36)
-			{
-				Direction = ParameterDirection.Input,
-				Value = storeGuid.ToString()
-			},
-			new("?CartGuid", MySqlDbType.VarChar, 36)
-			{
-				Direction = ParameterDirection.Input,
-				Value = cartGuid.ToString()
-			},
-			new("?RawResponse", MySqlDbType.Text)
-			{
-				Direction = ParameterDirection.Input,
-				Value = rawResponse
-			},
-			new("?ResponseCode", MySqlDbType.VarChar, 1)
-			{
-				Direction = ParameterDirection.Input,
-				Value = responseCode
-			},
-			new("?ResponseReasonCode", MySqlDbType.VarChar, 20)
-			{
-				Direction = ParameterDirection.Input,
-				Value = responseReasonCode
-			},
-			new("?Reason", MySqlDbType.Text)
-			{
-				Direction = ParameterDirection.Input,
-				Value = reason
-			},
-			new("?AvsCode", MySqlDbType.VarChar, 50)
-			{
-				Direction = ParameterDirection.Input,
-				Value = avsCode
-			},
-			new("?CcvCode", MySqlDbType.VarChar, 1)
-			{
-				Direction = ParameterDirection.Input,
-				Value = ccvCode
-			},
-			new("?CavCode", MySqlDbType.VarChar, 1)
-			{
-				Direction = ParameterDirection.Input,
-				Value = cavCode
-			},
-			new("?TransactionId", MySqlDbType.VarChar, 50)
-			{
-				Direction = ParameterDirection.Input,
-				Value = transactionId
-			},
-			new("?TransactionType", MySqlDbType.VarChar, 50)
-			{
-				Direction = ParameterDirection.Input,
-				Value = transactionType
-			},
-			new("?Method", MySqlDbType.VarChar, 20)
-			{
-				Direction = ParameterDirection.Input,
-				Value = method
-			},
-			new("?AuthCode", MySqlDbType.VarChar, 50)
-			{
-				Direction = ParameterDirection.Input,
-				Value = authCode
-			},
-			new("?Amount", MySqlDbType.Decimal)
-			{
-				Direction = ParameterDirection.Input,
-				Value = amount
-			},
-			new("?Tax", MySqlDbType.Decimal)
-			{
-				Direction = ParameterDirection.Input,
-				Value = tax
-			},
-			new("?Duty", MySqlDbType.Decimal)
-			{
-				Direction = ParameterDirection.Input,
-				Value = duty
-			},
-			new("?Freight", MySqlDbType.Decimal)
-			{
-				Direction = ParameterDirection.Input,
-				Value = freight
-			}
+			new("?RowGuid", MySqlDbType.VarChar, 36) { Value = rowGuid.ToString() },
+			new("?SiteGuid", MySqlDbType.VarChar, 36) { Value = siteGuid.ToString() },
+			new("?UserGuid", MySqlDbType.VarChar, 36) { Value = userGuid.ToString() },
+			new("?StoreGuid", MySqlDbType.VarChar, 36) { Value = storeGuid.ToString() },
+			new("?CartGuid", MySqlDbType.VarChar, 36) { Value = cartGuid.ToString() },
+			new("?RawResponse", MySqlDbType.Text) { Value = rawResponse },
+			new("?ResponseCode", MySqlDbType.VarChar, 1) { Value = responseCode },
+			new("?ResponseReasonCode", MySqlDbType.VarChar, 20) { Value = responseReasonCode },
+			new("?Reason", MySqlDbType.Text) { Value = reason },
+			new("?AvsCode", MySqlDbType.VarChar, 50) { Value = avsCode },
+			new("?CcvCode", MySqlDbType.VarChar, 1) { Value = ccvCode },
+			new("?CavCode", MySqlDbType.VarChar, 1) { Value = cavCode },
+			new("?TransactionId", MySqlDbType.VarChar, 50) { Value = transactionId },
+			new("?TransactionType", MySqlDbType.VarChar, 50) { Value = transactionType },
+			new("?Method", MySqlDbType.VarChar, 20) { Value = method },
+			new("?AuthCode", MySqlDbType.VarChar, 50) { Value = authCode },
+			new("?Amount", MySqlDbType.Decimal) { Value = amount },
+			new("?Tax", MySqlDbType.Decimal) { Value = tax },
+			new("?Duty", MySqlDbType.Decimal) { Value = duty },
+			new("?Freight", MySqlDbType.Decimal) { Value = freight }
 		};
 
-		int rowsAffected = CommandHelper.ExecuteNonQuery(
-			ConnectionString.GetWriteConnectionString(),
-			sqlCommand.ToString(),
-			arParams);
+		int rowsAffected = CommandHelper.ExecuteNonQuery(ConnectionString.GetWrite(), sqlCommand, arParams);
 
 		return rowsAffected > -1;
-
 	}
 
 	/// <summary>
@@ -434,27 +240,13 @@ WHERE
 	/// <returns>bool</returns>
 	public static bool Delete(Guid rowGuid)
 	{
-		string sqlCommand = @"
-DELETE FROM 
-	mp_AuthorizeNetLog 
-WHERE 
-	RowGuid = ?RowGuid;";
+		string sqlCommand = @"DELETE FROM mp_AuthorizeNetLog WHERE RowGuid = ?RowGuid;";
 
-		var arParams = new List<MySqlParameter>
-		{
-			new("?RowGuid", MySqlDbType.VarChar, 36)
-			{
-				Direction = ParameterDirection.Input,
-				Value = rowGuid.ToString()
-			}
-		};
+		var arParams = new MySqlParameter("?RowGuid", MySqlDbType.VarChar, 36) { Value = rowGuid.ToString() };
 
-		int rowsAffected = CommandHelper.ExecuteNonQuery(
-			ConnectionString.GetWriteConnectionString(),
-			sqlCommand.ToString(),
-			arParams);
+		int rowsAffected = CommandHelper.ExecuteNonQuery(ConnectionString.GetWrite(), sqlCommand, arParams);
+
 		return rowsAffected > 0;
-
 	}
 
 	/// <summary>
@@ -463,28 +255,11 @@ WHERE
 	/// <param name="rowGuid"> rowGuid </param>
 	public static IDataReader GetOne(Guid rowGuid)
 	{
-		string sqlCommand = @"
-SELECT  
-	* 
-FROM	
-	mp_AuthorizeNetLog 
-WHERE 
-	RowGuid = ?RowGuid;";
+		string sqlCommand = @"SELECT * FROM mp_AuthorizeNetLog WHERE RowGuid = ?RowGuid;";
 
-		var arParams = new List<MySqlParameter>
-		{
-			new("?RowGuid", MySqlDbType.VarChar, 36)
-			{
-				Direction = ParameterDirection.Input,
-				Value = rowGuid.ToString()
-			}
-		};
+		var param = new MySqlParameter("?RowGuid", MySqlDbType.VarChar, 36) { Value = rowGuid.ToString() };
 
-		return CommandHelper.ExecuteReader(
-			ConnectionString.GetReadConnectionString(),
-			sqlCommand.ToString(),
-			arParams);
-
+		return CommandHelper.ExecuteReader(ConnectionString.GetRead(), sqlCommand, param);
 	}
 
 	/// <summary>
@@ -493,63 +268,22 @@ WHERE
 	/// <param name="rowGuid"> rowGuid </param>
 	public static IDataReader GetByCart(Guid cartGuid)
 	{
-		string sqlCommand = @"
-SELECT  
-	*
-FROM	
-	mp_AuthorizeNetLog
-WHERE
-	CartGuid = ?CartGuid
-ORDER BY 
-	CreatedUtc;";
-		
-		var arParams = new List<MySqlParameter>
-		{
-			new("?CartGuid", MySqlDbType.VarChar, 36)
-			{
-				Direction = ParameterDirection.Input,
-				Value = cartGuid.ToString()
-			}
-		};
+		string sqlCommand = @"SELECT * FROM mp_AuthorizeNetLog WHERE CartGuid = ?CartGuid ORDER BY CreatedUtc;";
 
-		return CommandHelper.ExecuteReader(
-			ConnectionString.GetReadConnectionString(),
-			sqlCommand.ToString(),
-			arParams);
+		var param = new MySqlParameter("?CartGuid", MySqlDbType.VarChar, 36) { Value = cartGuid.ToString() };
 
+		return CommandHelper.ExecuteReader(ConnectionString.GetRead(), sqlCommand, param);
 	}
 
 	/// <summary>
 	/// Gets an IDataReader with all rows in the mp_AuthorizeNetLog table.
 	/// </summary>
-	public static IDataReader GetAll()
-	{
-		string sqlCommand = @"
-SELECT  
-	* 
-FROM	
-	mp_AuthorizeNetLog;";
-
-		return CommandHelper.ExecuteReader(
-			ConnectionString.GetReadConnectionString(),
-			sqlCommand.ToString());
-	}
+	public static IDataReader GetAll() => CommandHelper.ExecuteReader(ConnectionString.GetRead(), "SELECT * FROM mp_AuthorizeNetLog;");
 
 	/// <summary>
 	/// Gets a count of rows in the mp_AuthorizeNetLog table.
 	/// </summary>
-	public static int GetCount()
-	{
-		string sqlCommand = @"
-SELECT  
-	Count(*) 
-FROM	
-	mp_AuthorizeNetLog;";
-
-		return Convert.ToInt32(CommandHelper.ExecuteScalar(
-			ConnectionString.GetReadConnectionString(),
-			sqlCommand.ToString()));
-	}
+	public static int GetCount() => Convert.ToInt32(CommandHelper.ExecuteScalar(ConnectionString.GetRead(), "SELECT Count(*) FROM mp_AuthorizeNetLog;"));
 
 	/// <summary>
 	/// Gets a page of data from the mp_AuthorizeNetLog table.
@@ -557,10 +291,7 @@ FROM
 	/// <param name="pageNumber">The page number.</param>
 	/// <param name="pageSize">Size of the page.</param>
 	/// <param name="totalPages">total pages</param>
-	public static IDataReader GetPage(
-		int pageNumber,
-		int pageSize,
-		out int totalPages)
+	public static IDataReader GetPage(int pageNumber, int pageSize, out int totalPages)
 	{
 		int pageLowerBound = (pageSize * pageNumber) - pageSize;
 		totalPages = 1;
@@ -574,39 +305,21 @@ FROM
 		}
 		else
 		{
-			int remainder;
-			Math.DivRem(totalRows, pageSize, out remainder);
+			Math.DivRem(totalRows, pageSize, out int remainder);
 			if (remainder > 0)
 			{
 				totalPages += 1;
 			}
 		}
 
-		string sqlCommand = @"
-SELECT	
-	* 
-FROM	
-	mp_AuthorizeNetLog  
-LIMIT 
-	?Offset, 
-	?PageSize;";
+		string sqlCommand = @"SELECT * FROM	mp_AuthorizeNetLog LIMIT ?Offset, ?PageSize;";
 
 		MySqlParameter[] arParams =
 		[
-			new MySqlParameter("?Offset", MySqlDbType.Int32)
-			{
-				Direction = ParameterDirection.Input,
-				Value = pageLowerBound
-			},
-			new MySqlParameter("?PageSize", MySqlDbType.Int32)
-			{
-				Direction = ParameterDirection.Input,
-				Value = pageSize
-			},
+			new MySqlParameter("?Offset", MySqlDbType.Int32) { Value = pageLowerBound },
+			new MySqlParameter("?PageSize", MySqlDbType.Int32) { Value = pageSize },
 		];
-		return CommandHelper.ExecuteReader(
-			ConnectionString.GetReadConnectionString(),
-			sqlCommand.ToString(),
-			arParams);
+
+		return CommandHelper.ExecuteReader(ConnectionString.GetRead(), sqlCommand, arParams);
 	}
 }
